Feature ETP-4708: Extract ChildRowInlineEditor from DetailView (Ola 3)
First Ola 3 component extraction, chosen by derivation rather than by eye.

Twelve test files read DetailView.jsx as TEXT and regex its source: 39 such
assertions currently match. Moving any region they pin breaks them — behaviour
identical, test red, R1 violated. Matching every pin regex against every
candidate JSX region found six regions that break none. Of those, this one had
the best size-to-coupling ratio: 69 lines, 12 real props.

The component is a leaf — it owns no state, every value arrives as a prop, and
its PATCH/DELETE handlers are unchanged, which is what makes the move
behaviour-neutral.

It gets its own file rather than joining detailViewHelpers.jsx. That was the
first attempt, and the file-lines ratchet refused to raise the destination's
baseline (1743 -> 1832). It was right: piling a component into a 1.7k-line way
station trades one oversized file for another, against §9.2's ~600-line target.
The new file is tracked in this commit, before any baseline moved.

DetailView.jsx 3458 -> 3404. Verified: contract-ui vitest 2520/2520, identical
to the pre-extraction baseline.
